Fixed time formatter #9388
